Configure UI Kitten, and change Stack navigators location and import method
