+++ package.json
@@ -21,6 +21,7 @@
     ]
   },
   "dependencies": {
+    "@eva-design/eva": "^2.0.0",
     "@react-native-community/masked-view": "0.1.5",
     "@react-navigation/native": "^5.0.7",
     "@react-navigation/stack": "^5.0.8",
@@ -35,9 +36,9 @@
     "react-native-reanimated": "~1.4.0",
     "react-native-safe-area-context": "0.6.0",
     "react-native-screens": "2.0.0-alpha.12",
+    "react-native-svg": "^9.13.3",
     "react-native-swipe-gestures": "^1.0.5",
-    "react-native-web": "~0.11.7",
-    "react-native-svg": "9.13.3"
+    "react-native-web": "~0.11.7"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
